requirements.txt
amqp==5.3.1
asgiref==3.11.0
billiard==4.2.3
celery==5.5.3
channels==4.3.2
channels_redis==4.3.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
diff-match-patch==20241021
Django==5.2.3
django-ckeditor==6.7.3
django-ckeditor-5==0.2.18
django-cors-headers==4.9.0
django-filter==25.1
django-graphql-jwt==0.4.0
django-import-export==4.3.14
django-js-asset==3.1.2
django-redis==6.0.0
django-widget-tweaks==1.5.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
graphene==3.4.3
graphene-django==3.2.3
graphene-file-upload==1.3.0
graphql-core==3.2.6
graphql-relay==3.2.0
h11==0.16.0
kombu==5.5.4
lia-web==0.2.3
msgpack==1.1.2
packaging==25.0
pillow==11.2.1
promise==2.3
prompt_toolkit==3.0.52
psycopg==3.2.13
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dateutil==2.9.0.post0
redis==7.1.0
six==1.17.0
sqlparse==0.5.3
strawberry-graphql==0.287.0
strawberry-graphql-django==0.67.2
tablib==3.9.0
text-unidecode==1.3
typing_extensions==4.14.0
tzdata==2025.2
uvicorn==0.38.0
vine==5.1.0
wcwidth==0.2.14
